feat: implement lifecycleStore for managing AI instance lifecycle and refactor useLifecycle to utilize it
